split files and added interpolation trait
